Move stuff to DCIPUtils.py
