--- SimPEG/DCIP/DCIPUtils/__init__.py
@@ -1,1 +1,0 @@
-from readUBC_DC3Dobs import *
