+++ SimPEG/DCIP/__init__.py
@@ -1,4 +1,4 @@
 from BaseDC import *
 from BaseIP import *
-# from DCIPUtils import *
+from DCIPUtils import *
 import Utils
